SNOMED Code Usage in Primary Care in England

Yearly summary of code usage from 1st August 2011 to 31st July 2023.

Usage

snomed_usage

Format

snomed_usage

A data frame with 1,366,513 rows and 6 columns:

start_date

Start date of code usage count

end_date

End date of code usage count

snomed_concept_id

SNOMED Concept ID

usage

Yearly summary of code usage. Note that counts are rounded to the nearest 10. Counts of 5 or below are displayed as 5.

active_at_start

Specifying whether code was active at the start date.

active_at_end

Specifying whether code was active at the end date.

description

Description of SNOMED Concept ID

Source

https://digital.nhs.uk/data-and-information/publications/statistical/mi-snomed-code-usage-in-primary-care